debug backend delete message

+++ backend/util/db_queries.js
@@ -3,22 +3,22 @@ const fs = require('fs');
 
 // récupération des infos  de la ligne selon l'id du message
 exports.get_message_by_id = (id, callback) => {
-    connection.query('SELECT * FROM topic_messages WHERE tm_id = ?', [id], (err,results) => {
+    connection.query('SELECT * FROM topic_messages WHERE tm_id = ?', [id], (err, results) => {
         if (err) {
-            callback(err,null);
+            callback(err, null);
         } else {
-            callback(null,results);
+            callback(null, results);
         }
     });
 }
 
 // récupération de l'image selon l'id du message
 exports.get_picture_url_by_id = (id, callback) => {
-    connection.query('SELECT tm_picture_url FROM topic_messages WHERE tm_id = ?', [id], (err,results) => {
+    connection.query('SELECT tm_picture_url FROM topic_messages WHERE tm_id = ?', [id], (err, results) => {
         if (err) {
-            callback(err,null);
+            callback(err, null);
         } else {
-            callback(null,results);
+            callback(null, results);
         }
     });
 }
@@ -28,14 +28,16 @@ exports.delete_child_images_by_parent_id = (parent_id, callback) => {
     connection.query('SELECT tm_picture_url FROM topic_messages WHERE tm_parent = ?', [parent_id], (err, results) => {
         console.log(results);
         if (err) { // gestion de l'erreur
-            callback(err,null);
-        } else { 
+            callback(err, null);
+        } else {
             for (let index = 0; index < results.length; index++) { // boucle for pour récupérer la totalité des images dont le parent = 0
                 const element = results[index];
                 if (element.tm_picture_url) { // si on obtient un résultat dans la boucle : il existe au moins une image
                     try {
-                        //FileSystem : suppression des images/liens du "filesystem"
-                        fs.unlinkSync(element.tm_picture_url) 
+                        if (fs.existsSync(element.tm_picture_url)) { // vérification si le fichier existe
+                            //FileSystem : suppression des images/liens du "filesystem"
+                            fs.unlinkSync(element.tm_picture_url)
+                        }
                     } catch (err) { // gestion de l'erreur
                         console.log(err);
                     }
